requirements.txt
streamlit==1.29.0
langchain==0.1.0
chromadb==0.4.18
pypdf2==3.0.1
sentence-transformers==2.2.2
python-dotenv==1.0.0
tiktoken==0.5.2
pandas==2.1.4
numpy==1.24.3
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0